# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/spotify/artist": {
            "get": {
                "tags": [
                    "Spotify Api"
                ],
                "summary": "Get Artist",
                "description": "Required => token, artistId",
                "operationId": "d7b9f45fb673252ccc04b76808099510",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/spotify/artist/genres": {
            "get": {
                "tags": [
                    "Spotify Api"
                ],
                "summary": "Get Artist Genres",
                "description": "Required => token, artistId",
                "operationId": "7fed825efdc13215b9a717f3e36e894e",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/spotify/artist/albums": {
            "get": {
                "tags": [
                    "Spotify Api"
                ],
                "summary": "Get Artist Albums",
                "description": "Required => token, artistId",
                "operationId": "4a768261ef06f2cb8bce018c7e03a789",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "market",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/spotify/album/tracks": {
            "get": {
                "tags": [
                    "Spotify Api"
                ],
                "summary": "Get Album's Track",
                "description": "Required => token, artistId, albumId",
                "operationId": "b09e083b4c694fa55e64edb1178e6cdf",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "albumId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "market",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/spotify/artist/tracks": {
            "get": {
                "tags": [
                    "Spotify Api"
                ],
                "summary": "Get Artist's Track",
                "description": "Required => token, artistId\n\n    After fetching an album, there is a 2-second wait for the next album.",
                "operationId": "61c9cfd8a3014c54235b614515921e18",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "market",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}